+++ src/application/grpc/health.rs
@@ -1,0 +1,75 @@
+use crate::application::Application;
+use crate::prostgen::cosmos::base::tendermint::v1beta1::{
+    service_server::Service as HealthService, GetBlockByHeightRequest, GetBlockByHeightResponse,
+    GetLatestBlockRequest, GetLatestBlockResponse, GetLatestValidatorSetRequest,
+    GetLatestValidatorSetResponse, GetNodeInfoRequest, GetNodeInfoResponse, GetSyncingRequest,
+    GetSyncingResponse, GetValidatorSetByHeightRequest, GetValidatorSetByHeightResponse,
+    Module as VersionInfoModule, VersionInfo,
+};
+use crate::store::ProvableStore;
+
+use tendermint_proto::p2p::DefaultNodeInfo;
+use tonic::{Request, Response, Status};
+use tracing::debug;
+
+#[tonic::async_trait]
+impl<S: ProvableStore + 'static> HealthService for Application<S> {
+    async fn get_node_info(
+        &self,
+        _request: Request<GetNodeInfoRequest>,
+    ) -> Result<Response<GetNodeInfoResponse>, Status> {
+        debug!("Got node info request");
+
+        // TODO(hu55a1n1): generate below info using build script
+        Ok(Response::new(GetNodeInfoResponse {
+            default_node_info: Some(DefaultNodeInfo::default()),
+            application_version: Some(VersionInfo {
+                name: "basecoin-rs".to_string(),
+                app_name: "basecoind".to_string(),
+                version: "0.1.0".to_string(),
+                git_commit: "209afef7e99ebcb814b25b6738d033aa5e1a932c".to_string(),
+                build_deps: vec![VersionInfoModule {
+                    path: "github.com/cosmos/cosmos-sdk".to_string(),
+                    version: "v0.43.0".to_string(),
+                    sum: "h1:ps1QWfvaX6VLNcykA7wzfii/5IwBfYgTIik6NOVDq/c=".to_string(),
+                }],
+                ..VersionInfo::default()
+            }),
+        }))
+    }
+
+    async fn get_syncing(
+        &self,
+        _request: Request<GetSyncingRequest>,
+    ) -> Result<Response<GetSyncingResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_latest_block(
+        &self,
+        _request: Request<GetLatestBlockRequest>,
+    ) -> Result<Response<GetLatestBlockResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_block_by_height(
+        &self,
+        _request: Request<GetBlockByHeightRequest>,
+    ) -> Result<Response<GetBlockByHeightResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_latest_validator_set(
+        &self,
+        _request: Request<GetLatestValidatorSetRequest>,
+    ) -> Result<Response<GetLatestValidatorSetResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_validator_set_by_height(
+        &self,
+        _request: Request<GetValidatorSetByHeightRequest>,
+    ) -> Result<Response<GetValidatorSetByHeightResponse>, Status> {
+        unimplemented!()
+    }
+}
