+++ src/application/grpc/mod.rs
@@ -1,0 +1,2 @@
+mod health;
+mod tx;
